Fixed some compiler warnings

+++ src/telescope/mount/Mount.command.cpp
@@ -68,8 +68,8 @@ bool Mount::command(char *reply, char *command, char *parameter, bool *supressFr
           case '1': convert.doubleToDms(reply, radToDeg(axis2.getInstrumentCoordinate()), true, true, PM_HIGH); break; 
           case '2': sprintF(reply, "%0.6f", radToDeg(axis1.getInstrumentCoordinate())); break;
           case '3': sprintF(reply, "%0.6f", radToDeg(axis2.getInstrumentCoordinate())); break;
-          case '4': sprintf(reply, "%d", axis1.motor->getEncoderCount()); break;
-          case '5': sprintf(reply, "%d", axis2.motor->getEncoderCount()); break;
+          case '4': sprintf(reply, "%ld", (long)axis1.motor->getEncoderCount()); break;
+          case '5': sprintf(reply, "%ld", (long)axis2.motor->getEncoderCount()); break;
           default:  *numericReply = true; *commandError = CE_CMD_UNKNOWN;
         }
       } else
